updates
proposed another way to plot the carbon and nitrogen weight

corrected predictions in predict file for reprod rate, nitrogen mass

this is model sbp so do not multiply with shape correction factor that is for apb (when we try that next :-)

i am not 100% sure about the carbon and nitrogen weights, we need to check. But.... n_CC is 1 by definition that is why it did not figure in the prediction for carbon weight.

n_NV on the other hand, is the amounts of N per C in structure and I added that to the prediction for the carbon weight

in pars init ia dded a line for you to see how to overwrite the default chemical indices and I put n_NV to 0.2 just so you can play a bit with the value. Think of some generalized mixture of carbohydrates, lipids and proteins. and what could be the amount of N atmoms to C atoms in that mixture (which is strongly homeostatic).

the weakness in what I did is that i think the formula assumes n_NV = n_NE (as only the chemcical indice for nitrogen in reserve appear in the formula ), with the overwriting in the pars init be aware that now n_NE is different from n_NV  and the formula migth not be super correct. we can look at that tomorrow and beyond.